added I2C driver
